refactor: move accounting::ServerState(Import) from lib to wire
